Изменение структуры каталогов store
