+++ src/index.js
@@ -10,7 +10,7 @@ import {generateMockReviews} from "./mocks/reviews";
 import {createAPI} from "./services/api";
 import {fetchFilms, fetchPromoFilm} from "./store/api-actions";
 import {redirect} from "./store/middlewares/redirect";
-import rootReducer from "./store/reducers/root-reducer";
+import reducer from "./store/reducer";
 
 const api = createAPI(()=>{});
 
@@ -20,7 +20,7 @@ const composeWithEnhancers = composeWithDevTools({
   trace: true,
 });
 
-const store = createStore(rootReducer,
+const store = createStore(reducer,
     composeWithEnhancers(
         applyMiddleware(thunk.withExtraArgument(api)),
         applyMiddleware(redirect)));
