+++ src/store/reducers/catalog.js
@@ -1,6 +1,6 @@
-import {ALL_GENRES_FILTER, SHOWN_FILMS_INITIAL_COUNT} from "../../../const";
-import {extend} from "../../../utils/common";
-import {ActionType} from "../../action";
+import {ALL_GENRES_FILTER, SHOWN_FILMS_INITIAL_COUNT} from "../../const";
+import {extend} from "../../utils/common";
+import {ActionType} from "../action";
 
 const initialState = {
   currentGenre: ALL_GENRES_FILTER,
@@ -8,7 +8,7 @@ const initialState = {
 
 };
 
-export const filmCatalog = (state = initialState, action) => {
+export const catalog = (state = initialState, action) => {
   switch (action.type) {
 
     case ActionType.CHANGE_CURRENT_GENRE:
