+++ src/store/reducers/user.js
@@ -1,6 +1,5 @@
-import {AuthorizationStatus} from "../../../const";
-import {ActionType} from "../../action";
-
+import {AuthorizationStatus} from "../../const";
+import {ActionType} from "../action";
 const initialState = {
   authorizationStatus: AuthorizationStatus.NO_AUTH,
 };
